IMP: Put hero in its own component
